modules/digitalocean/master: add droplet_id to floating_ip resource
Fixes #1, #2.

Remove `local-exec` provisioner from master_node. No need to do this in
a posthook, Terraform will wait for the resource to be provisioned
before assigning the Droplet to the Floating IP.

There are no dependencies on this to be pre-computed and used in a
template file that would block them from using the Terraform lifecycle &
dependency built-ins.

Disable `depends_on` for `tectonic_assets` and `kubeconfig` null
resources temporarily.